fourth commit

+++ base/views.py
@@ -1,17 +1,22 @@
 from django.shortcuts import render,redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login,logout
-from .models import Room,Topic
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room,Topic,Message
 from .forms import RoomForm
 
 # Create your views here.
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method =='POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username = username)
@@ -24,9 +29,28 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or password does not exist!')
             
-    context={}   
+    context={'page':page}   
     return render(request,'base/login_registration.html', context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+
+    form= UserCreationForm()
+    if request.method =='POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid:
+            user =form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"An error occurred during registration")
+    
+    return render(request, 'base/login_registration.html', {'form':form} )
 def home(request):
     #to search by topic name
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -41,12 +65,24 @@ def home(request):
 
 def room(request,pk):
     room = Room.objects.get(id = pk)
-    # to go to the room
-    
-    context ={'room': room}
+    #room messages
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+   
+    if request.method =='POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
             
+        )
+        room.participants.add(request.user)
+        return redirect('room',pk = room.id)
+    
+    context ={'room': room, 'room_messages':room_messages,'participants':participants}       
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     
@@ -60,9 +96,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
+    
     if request.method =='POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -72,9 +112,23 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj': room})
+
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    
+    if request.user != message.user:
+        return HttpResponse("You are not allowed here!")
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': message})
